Tank Program
This is a fairly simple Python program using the Pygame library.  There are two tanks at opposite end of the battle field. The yellow tank is currently controlled by the arrow keys.  Left and right rotate the turret. The tank always moves in the direction of the turret.  The up arrow increases the speed. The down arrow slows down the speed.  The end key does the shoot.  At this time the red tank keys are not active, so it is a single player game.

There is a collection of robot tanks that are randomly placed on the field.  They lock on to another tank and shoot to kill. Their range is more limited than the controlled tank. But they are relentless.  The object at this time is to move the yellow tank through the field and take out the red tank.  To spice things up, there are mines placed randomly on the field. They are mainly invisible but show up briefly at the beginning as blue boxes and then periodically show briefly. So their location must be quickly noted. 

Several things needed now: